Django-signals_orm-0x04/pyproject.toml
[project]
name = "django-signals-orm-0x04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
]
